Refatora holistic-sheet em hooks/core e seção de toolbar
