+++ components/ui-grid/sections/drawers.tsx
@@ -1,0 +1,5 @@
+import { ReactNode } from "react";
+
+export function GridDrawersSection({ children }: { children: ReactNode }) {
+  return <>{children}</>;
+}
