+++ components/ui-grid/sections/sidepanels.tsx
@@ -1,0 +1,5 @@
+import { ReactNode } from "react";
+
+export function GridSidePanelsSection({ children }: { children: ReactNode }) {
+  return <>{children}</>;
+}
