+++ components/ui-grid/sections/table-body.tsx
@@ -1,0 +1,5 @@
+import { ReactNode } from "react";
+
+export function GridTableBodySection({ children }: { children: ReactNode }) {
+  return <>{children}</>;
+}
